+++ lib/src/commands/generate/generate.dart
@@ -22,7 +22,6 @@ class GenerateCommand extends Command<int> {
     addSubcommand(PageCommand(logger));
     addSubcommand(SourceCommand(logger));
     addSubcommand(RepositoryCommand(logger));
-    addSubcommand(EnvCommand(logger));
     addSubcommand(ModelCommand(logger));
     addSubcommand(AssetsCommand(logger));
   }