+++ lib/src/commands/generate/templates/graphql/feature/graphql_feature_bundle.dart
@@ -3,4 +3,113 @@
 
 import 'package:mason/mason.dart';
 
-final graphqlFeatureBundle = MasonBundle.fromJson(<String, dynamic>{"files":[{"path":"{{name.snakeCase()}}/data/graphql/{{name.snakeCase()}}_mutations.dart","data":"LyoKICogQ29weXJpZ2h0IChjKSAyMDIyCiAqIEF1dGhvcjogS2lzaG9yIE1haW5hbGkKICogQ29tcGFueTogRUIgUGVhcmxzCiAqLwoKY2xhc3Mge3tuYW1lLnBhc2NhbENhc2UoKX19TXV0YXRpb25zIHsKICBjb25zdCB7e25hbWUucGFzY2FsQ2FzZSgpfX1NdXRhdGlvbnMuXygpOwp9Cg==","type":"text"},{"path":"{{name.snakeCase()}}/data/graphql/{{name.snakeCase()}}_queries.dart","data":"LyoKICogQ29weXJpZ2h0IChjKSAyMDIyCiAqIEF1dGhvcjogS2lzaG9yIE1haW5hbGkKICogQ29tcGFueTogRUIgUGVhcmxzCiAqLwoKY2xhc3Mge3tuYW1lLnBhc2NhbENhc2UoKX19UXVlcmllcyB7CmNvbnN0IHt7bmFtZS5wYXNjYWxDYXNlKCl9fVF1ZXJpZXMuXygpOwp9Cg==","type":"text"},{"path":"{{name.snakeCase()}}/data/models/{{name.snakeCase()}}_model.dart","data":"LyoKICogQ29weXJpZ2h0IChjKSAyMDIyCiAqIEF1dGhvcjogS2lzaG9yIE1haW5hbGkKICogQ29tcGFueTogRUIgUGVhcmxzCiAqLwoKaW1wb3J0ICdwYWNrYWdlOmZyZWV6ZWRfYW5ub3RhdGlvbi9mcmVlemVkX2Fubm90YXRpb24uZGFydCc7CgpwYXJ0ICd7e25hbWUuc25ha2VDYXNlKCl9fV9tb2RlbC5mcmVlemVkLmRhcnQnOwoKcGFydCAne3tuYW1lLnNuYWtlQ2FzZSgpfX1fbW9kZWwuZy5kYXJ0JzsKCkBmcmVlemVkCmNsYXNzIHt7bmFtZS5wYXNjYWxDYXNlKCl9fU1vZGVsIHdpdGggXyR7e25hbWUucGFzY2FsQ2FzZSgpfX1Nb2RlbCB7CmNvbnN0IHt7bmFtZS5wYXNjYWxDYXNlKCl9fU1vZGVsLl8oKTsKCmNvbnN0IGZhY3Rvcnkge3tuYW1lLnBhc2NhbENhc2UoKX19TW9kZWwoKSA9IF97e25hbWUucGFzY2FsQ2FzZSgpfX1Nb2RlbDsKCmZhY3Rvcnkge3tuYW1lLnBhc2NhbENhc2UoKX19TW9kZWwuZnJvbUpzb24oTWFwPFN0cmluZywgZHluYW1pYz4ganNvbikgPT4gXyR7e25hbWUucGFzY2FsQ2FzZSgpfX1Nb2RlbEZyb21Kc29uKGpzb24pOwoKfQo=","type":"text"},{"path":"{{name.snakeCase()}}/data/repository/{{name.snakeCase()}}_repository_impl.dart","data":"LyoKICogQ29weXJpZ2h0IChjKSAyMDIyCiAqIEF1dGhvcjogS2lzaG9yIE1haW5hbGkKICogQ29tcGFueTogRUIgUGVhcmxzCiAqLwoKaW1wb3J0ICdkYXJ0OmFzeW5jJzsKCmltcG9ydCAncGFja2FnZTp7e3BhY2thZ2VfbmFtZS5zbmFrZUNhc2UoKX19L3NyYy9jb3JlL2Jhc2UvYmFzZV9yZXBvc2l0b3J5LmRhcnQnOwppbXBvcnQgJ3BhY2thZ2U6aW5qZWN0YWJsZS9pbmplY3RhYmxlLmRhcnQnOwoKaW1wb3J0ICcuLi8uLi9kb21haW4vcmVwb3NpdG9yeS97e25hbWUuc25ha2VDYXNlKCl9fV9yZXBvc2l0b3J5LmRhcnQnOwppbXBvcnQgJy4uL3NvdXJjZS97e25hbWUuc25ha2VDYXNlKCl9fV9yZW1vdGVfc291cmNlLmRhcnQnOwoKQExhenlTaW5nbGV0b24oYXM6IHt7bmFtZS5wYXNjYWxDYXNlKCl9fVJlcG9zaXRvcnkpCmNsYXNzIHt7bmFtZS5wYXNjYWxDYXNlKCl9fVJlcG9zaXRvcnlJbXBsIGV4dGVuZHMgQmFzZVJlcG9zaXRvcnkgaW1wbGVtZW50cyB7e25hbWUucGFzY2FsQ2FzZSgpfX1SZXBvc2l0b3J5IHsKe3tuYW1lLnBhc2NhbENhc2UoKX19UmVwb3NpdG9yeUltcGwodGhpcy5fcmVtb3RlU291cmNlLCBzdXBlci5uZXR3b3JrSW5mbyk7CgpmaW5hbCB7e25hbWUucGFzY2FsQ2FzZSgpfX1SZW1vdGVTb3VyY2UgX3JlbW90ZVNvdXJjZTsKCn0K","type":"text"},{"path":"{{name.snakeCase()}}/data/source/{{name.snakeCase()}}_remote_source.dart","data":"LyoKICogQ29weXJpZ2h0IChjKSAyMDIyCiAqIEF1dGhvcjogS2lzaG9yIE1haW5hbGkKICogQ29tcGFueTogRUIgUGVhcmxzCiAqLwoKaW1wb3J0ICdwYWNrYWdlOmluamVjdGFibGUvaW5qZWN0YWJsZS5kYXJ0JzsKaW1wb3J0ICdwYWNrYWdlOmdyYXBocWwvY2xpZW50LmRhcnQnOwoKaW1wb3J0ICdwYWNrYWdlOnt7cGFja2FnZV9uYW1lLnNuYWtlQ2FzZSgpfX0vc3JjL2NvcmUvYmFzZS9iYXNlX2dyYXBoX3JlbW90ZV9zb3VyY2UuZGFydCc7CgoKYWJzdHJhY3QgY2xhc3Mge3tuYW1lLnBhc2NhbENhc2UoKX19UmVtb3RlU291cmNlIHsKfQoKQExhenlTaW5nbGV0b24oYXM6IHt7bmFtZS5wYXNjYWxDYXNlKCl9fVJlbW90ZVNvdXJjZSkKY2xhc3Mge3tuYW1lLnBhc2NhbENhc2UoKX19UmVtb3RlU291cmNlSW1wbCBleHRlbmRzIEJhc2VHcmFwaFJlbW90ZVNvdXJjZSBpbXBsZW1lbnRzIHt7bmFtZS5wYXNjYWxDYXNlKCl9fVJlbW90ZVNvdXJjZSB7Cnt7bmFtZS5wYXNjYWxDYXNlKCl9fVJlbW90ZVNvdXJjZUltcGwoR3JhcGhRTENsaWVudCBzdXBlci5jbGllbnQpOwp9Cg==","type":"text"},{"path":"{{name.snakeCase()}}/domain/repository/{{name.snakeCase()}}_repository.dart","data":"LyoKICogQ29weXJpZ2h0IChjKSAyMDIyCiAqIEF1dGhvcjogS2lzaG9yIE1haW5hbGkKICogQ29tcGFueTogRUIgUGVhcmxzCiAqLwoKYWJzdHJhY3QgY2xhc3Mge3tuYW1lLnBhc2NhbENhc2UoKX19UmVwb3NpdG9yeSB7fQo=","type":"text"},{"path":"{{name.snakeCase()}}/presentation/pages/{{name.snakeCase()}}_page.dart","data":"LyoKICogQ29weXJpZ2h0IChjKSAyMDIyCiAqIEF1dGhvcjogS2lzaG9yIE1haW5hbGkKICogQ29tcGFueTogRUIgUGVhcmxzCiAqLwppbXBvcnQgJ3BhY2thZ2U6Zmx1dHRlci9tYXRlcmlhbC5kYXJ0JzsKe3sjc3RhdGV9fQpjbGFzcyB7e25hbWUucGFzY2FsQ2FzZSgpfX1QYWdlIGV4dGVuZHMgU3RhdGVmdWxXaWRnZXQgewogIGNvbnN0IHt7bmFtZS5wYXNjYWxDYXNlKCl9fVBhZ2Uoe0tleT8ga2V5fSkgOiBzdXBlcihrZXk6IGtleSk7CgogIEBvdmVycmlkZQogIF97e25hbWUucGFzY2FsQ2FzZSgpfX1QYWdlU3RhdGUgY3JlYXRlU3RhdGUoKSA9PiBfe3tuYW1lLnBhc2NhbENhc2UoKX19UGFnZVN0YXRlKCk7Cn0KCmNsYXNzIF97e25hbWUucGFzY2FsQ2FzZSgpfX1QYWdlU3RhdGUgZXh0ZW5kcyBTdGF0ZTx7e25hbWUucGFzY2FsQ2FzZSgpfX1QYWdlPiB7CiAgQG92ZXJyaWRlCiAgV2lkZ2V0IGJ1aWxkKEJ1aWxkQ29udGV4dCBjb250ZXh0KSB7CiAgICByZXR1cm4gU2NhZmZvbGQoKTsKICB9Cn0Ke3svc3RhdGV9fQoKe3tec3RhdGV9fQpjbGFzcyB7e25hbWUucGFzY2FsQ2FzZSgpfX1QYWdlIGV4dGVuZHMgU3RhdGVsZXNzV2lkZ2V0IHsKY29uc3Qge3tuYW1lLnBhc2NhbENhc2UoKX19UGFnZSh7c3VwZXIua2V5fSk7CgpAb3ZlcnJpZGUKV2lkZ2V0IGJ1aWxkKEJ1aWxkQ29udGV4dCBjb250ZXh0KSB7CnJldHVybiBTY2FmZm9sZCgpOwp9Cn0Ke3svc3RhdGV9fQ==","type":"text"}],"hooks":[{"path":"post_gen.dart","data":"aW1wb3J0ICdkYXJ0OmlvJzsKCmltcG9ydCAncGFja2FnZTptYXNvbi9tYXNvbi5kYXJ0JzsKCnZvaWQgcnVuKEhvb2tDb250ZXh0IGNvbnRleHQpIHsKICBmaW5hbCBmaWxlID0gRmlsZSgne3tuYW1lLnNuYWtlQ2FzZSgpfX0uZGFydCcpOwogIGlmICghZmlsZS5leGlzdHNTeW5jKCkpIGZpbGUuY3JlYXRlU3luYygpOwogIGZpbGUud3JpdGVBc1N0cmluZ1N5bmMoCiAgICAgICdleHBvcnQgXCdwcmVzZW50YXRpb24vcGFnZXMve3tuYW1lLnNuYWtlQ2FzZSgpfX1fcGFnZS5kYXJ0XCc7XG4nLAogICAgICBtb2RlOiBGaWxlTW9kZS5hcHBlbmQpOwp9Cg==","type":"text"},{"path":"pubspec.yaml","data":"bmFtZTogY2xlYW5fZmVhdHVyZV9ob29rcwoKZW52aXJvbm1lbnQ6CiAgc2RrOiAiPj0yLjE3LjAgPDMuMC4wIgoKZGVwZW5kZW5jaWVzOgogIG1hc29uOiBhbnkK","type":"text"}],"name":"graphql_feature","description":"A new brick created with the Mason CLI.","version":"0.1.0+1","environment":{"mason":">=0.1.0-dev <0.1.0"},"readme":{"path":"README.md","data":"IyBjbGVhbl9mZWF0dXJlCgpBIG5ldyBicmljayBjcmVhdGVkIHdpdGggdGhlIE1hc29uIENMSS4KCl9HZW5lcmF0ZWQgYnkgW21hc29uXVsxXSDwn6exXwoKIyMgR2V0dGluZyBTdGFydGVkIPCfmoAKClRoaXMgaXMgYSBzdGFydGluZyBwb2ludCBmb3IgYSBuZXcgYnJpY2suCkEgZmV3IHJlc291cmNlcyB0byBnZXQgeW91IHN0YXJ0ZWQgaWYgdGhpcyBpcyB5b3VyIGZpcnN0IGJyaWNrIHRlbXBsYXRlOgoKLSBbT2ZmaWNpYWwgTWFzb24gRG9jdW1lbnRhdGlvbl1bMl0KLSBbQ29kZSBnZW5lcmF0aW9uIHdpdGggTWFzb24gQmxvZ11bM10KLSBbVmVyeSBHb29kIExpdmVzdHJlYW06IEZlbGl4IEFuZ2Vsb3YgRGVtb3MgTWFzb25dWzRdCgpbMV06IGh0dHBzOi8vZ2l0aHViLmNvbS9mZWxhbmdlbC9tYXNvbgpbMl06IGh0dHBzOi8vZ2l0aHViLmNvbS9mZWxhbmdlbC9tYXNvbi90cmVlL21hc3Rlci9wYWNrYWdlcy9tYXNvbl9jbGkjcmVhZG1lClszXTogaHR0cHM6Ly92ZXJ5Z29vZC52ZW50dXJlcy9ibG9nL2NvZGUtZ2VuZXJhdGlvbi13aXRoLW1hc29uCls0XTogaHR0cHM6Ly95b3V0dS5iZS9HNFBUakE2dHBUVQo=","type":"text"},"changelog":{"path":"CHANGELOG.md","data":"IyAwLjEuMCsxCgotIFRPRE86IERlc2NyaWJlIGluaXRpYWwgcmVsZWFzZS4K","type":"text"},"license":{"path":"LICENSE","data":"VE9ETzogQWRkIHlvdXIgbGljZW5zZSBoZXJlLgo=","type":"text"},"vars":{"name":{"type":"string","description":"Your feature name","prompt":"What is your feature name?"},"useDio":{"type":"boolean","description":"client type","default":false,"prompt":"use dio?"},"package_name":{"type":"string","description":"Your package name","prompt":"What is your package name?"},"state":{"type":"boolean","description":"Page type","default":false,"prompt":"What is your page type <stateless,stateful>?"}}});
+final graphqlFeatureBundle = MasonBundle.fromJson(<String, dynamic>{
+  "files": [
+    {
+      "path":
+          "{{name.snakeCase()}}/data/graphql/{{name.snakeCase()}}_mutation.graphql",
+      "data":
+          "Cm11dGF0aW9uIHt7bmFtZS5wYXNjYWxDYXNlKCl9fU11dGF0aW9uKCkgewoKfQo=",
+      "type": "text"
+    },
+    {
+      "path":
+          "{{name.snakeCase()}}/data/graphql/{{name.snakeCase()}}_query.graphql",
+      "data": "CnF1ZXJ5IHt7bmFtZS5wYXNjYWxDYXNlKCl9fVF1ZXJ5IHsKfQo=",
+      "type": "text"
+    },
+    {
+      "path":
+          "{{name.snakeCase()}}/data/models/{{name.snakeCase()}}_model.dart",
+      "data":
+          "LyoKICogQ29weXJpZ2h0IChjKSAyMDIyCiAqIEF1dGhvcjogS2lzaG9yIE1haW5hbGkKICogQ29tcGFueTogRUIgUGVhcmxzCiAqLwoKaW1wb3J0ICdwYWNrYWdlOmZyZWV6ZWRfYW5ub3RhdGlvbi9mcmVlemVkX2Fubm90YXRpb24uZGFydCc7CgpwYXJ0ICd7e25hbWUuc25ha2VDYXNlKCl9fV9tb2RlbC5mcmVlemVkLmRhcnQnOwoKcGFydCAne3tuYW1lLnNuYWtlQ2FzZSgpfX1fbW9kZWwuZy5kYXJ0JzsKCkBmcmVlemVkCmNsYXNzIHt7bmFtZS5wYXNjYWxDYXNlKCl9fU1vZGVsIHdpdGggXyR7e25hbWUucGFzY2FsQ2FzZSgpfX1Nb2RlbCB7CmNvbnN0IHt7bmFtZS5wYXNjYWxDYXNlKCl9fU1vZGVsLl8oKTsKCmNvbnN0IGZhY3Rvcnkge3tuYW1lLnBhc2NhbENhc2UoKX19TW9kZWwoKSA9IF97e25hbWUucGFzY2FsQ2FzZSgpfX1Nb2RlbDsKCmZhY3Rvcnkge3tuYW1lLnBhc2NhbENhc2UoKX19TW9kZWwuZnJvbUpzb24oTWFwPFN0cmluZywgZHluYW1pYz4ganNvbikgPT4gXyR7e25hbWUucGFzY2FsQ2FzZSgpfX1Nb2RlbEZyb21Kc29uKGpzb24pOwoKfQo=",
+      "type": "text"
+    },
+    {
+      "path":
+          "{{name.snakeCase()}}/data/repository/{{name.snakeCase()}}_repository_impl.dart",
+      "data":
+          "LyoKICogQ29weXJpZ2h0IChjKSAyMDIyCiAqIEF1dGhvcjogS2lzaG9yIE1haW5hbGkKICogQ29tcGFueTogRUIgUGVhcmxzCiAqLwoKaW1wb3J0ICdkYXJ0OmFzeW5jJzsKCmltcG9ydCAncGFja2FnZTp7e3BhY2thZ2VfbmFtZS5zbmFrZUNhc2UoKX19L3NyYy9jb3JlL2Jhc2UvYmFzZV9yZXBvc2l0b3J5LmRhcnQnOwppbXBvcnQgJ3BhY2thZ2U6aW5qZWN0YWJsZS9pbmplY3RhYmxlLmRhcnQnOwoKaW1wb3J0ICcuLi8uLi9kb21haW4vcmVwb3NpdG9yeS97e25hbWUuc25ha2VDYXNlKCl9fV9yZXBvc2l0b3J5LmRhcnQnOwppbXBvcnQgJy4uL3NvdXJjZS97e25hbWUuc25ha2VDYXNlKCl9fV9yZW1vdGVfc291cmNlLmRhcnQnOwoKQExhenlTaW5nbGV0b24oYXM6IHt7bmFtZS5wYXNjYWxDYXNlKCl9fVJlcG9zaXRvcnkpCmNsYXNzIHt7bmFtZS5wYXNjYWxDYXNlKCl9fVJlcG9zaXRvcnlJbXBsIGV4dGVuZHMgQmFzZVJlcG9zaXRvcnkgaW1wbGVtZW50cyB7e25hbWUucGFzY2FsQ2FzZSgpfX1SZXBvc2l0b3J5IHsKe3tuYW1lLnBhc2NhbENhc2UoKX19UmVwb3NpdG9yeUltcGwodGhpcy5fcmVtb3RlU291cmNlKTsKCmZpbmFsIHt7bmFtZS5wYXNjYWxDYXNlKCl9fVJlbW90ZVNvdXJjZSBfcmVtb3RlU291cmNlOwoKfQo=",
+      "type": "text"
+    },
+    {
+      "path":
+          "{{name.snakeCase()}}/data/source/{{name.snakeCase()}}_remote_source.dart",
+      "data":
+          "LyoKICogQ29weXJpZ2h0IChjKSAyMDIyCiAqIEF1dGhvcjogS2lzaG9yIE1haW5hbGkKICogQ29tcGFueTogRUIgUGVhcmxzCiAqLwoKaW1wb3J0ICdwYWNrYWdlOmluamVjdGFibGUvaW5qZWN0YWJsZS5kYXJ0JzsKaW1wb3J0ICdwYWNrYWdlOmZlcnJ5L2ZlcnJ5LmRhcnQnOwppbXBvcnQgJ3BhY2thZ2U6e3twYWNrYWdlX25hbWUuc25ha2VDYXNlKCl9fS9zcmMvY29yZS9iYXNlL2Jhc2VfcmVtb3RlX3NvdXJjZS5kYXJ0JzsKCgphYnN0cmFjdCBjbGFzcyB7e25hbWUucGFzY2FsQ2FzZSgpfX1SZW1vdGVTb3VyY2Ugewp9CgpATGF6eVNpbmdsZXRvbihhczoge3tuYW1lLnBhc2NhbENhc2UoKX19UmVtb3RlU291cmNlKQpjbGFzcyB7e25hbWUucGFzY2FsQ2FzZSgpfX1SZW1vdGVTb3VyY2VJbXBsIGV4dGVuZHMgQmFzZVJlbW90ZVNvdXJjZSBpbXBsZW1lbnRzIHt7bmFtZS5wYXNjYWxDYXNlKCl9fVJlbW90ZVNvdXJjZSB7Cnt7bmFtZS5wYXNjYWxDYXNlKCl9fVJlbW90ZVNvdXJjZUltcGwoQ2xpZW50IHN1cGVyLmNsaWVudCk7Cn0K",
+      "type": "text"
+    },
+    {
+      "path":
+          "{{name.snakeCase()}}/domain/repository/{{name.snakeCase()}}_repository.dart",
+      "data":
+          "LyoKICogQ29weXJpZ2h0IChjKSAyMDIyCiAqIEF1dGhvcjogS2lzaG9yIE1haW5hbGkKICogQ29tcGFueTogRUIgUGVhcmxzCiAqLwoKYWJzdHJhY3QgY2xhc3Mge3tuYW1lLnBhc2NhbENhc2UoKX19UmVwb3NpdG9yeSB7fQo=",
+      "type": "text"
+    },
+    {
+      "path":
+          "{{name.snakeCase()}}/presentation/pages/{{name.snakeCase()}}_page.dart",
+      "data":
+          "LyoKICogQ29weXJpZ2h0IChjKSAyMDIyCiAqIEF1dGhvcjogS2lzaG9yIE1haW5hbGkKICogQ29tcGFueTogRUIgUGVhcmxzCiAqLwppbXBvcnQgJ3BhY2thZ2U6Zmx1dHRlci9tYXRlcmlhbC5kYXJ0JzsKe3sjc3RhdGV9fQpjbGFzcyB7e25hbWUucGFzY2FsQ2FzZSgpfX1QYWdlIGV4dGVuZHMgU3RhdGVmdWxXaWRnZXQgewogIGNvbnN0IHt7bmFtZS5wYXNjYWxDYXNlKCl9fVBhZ2Uoe0tleT8ga2V5fSkgOiBzdXBlcihrZXk6IGtleSk7CgogIEBvdmVycmlkZQogIF97e25hbWUucGFzY2FsQ2FzZSgpfX1QYWdlU3RhdGUgY3JlYXRlU3RhdGUoKSA9PiBfe3tuYW1lLnBhc2NhbENhc2UoKX19UGFnZVN0YXRlKCk7Cn0KCmNsYXNzIF97e25hbWUucGFzY2FsQ2FzZSgpfX1QYWdlU3RhdGUgZXh0ZW5kcyBTdGF0ZTx7e25hbWUucGFzY2FsQ2FzZSgpfX1QYWdlPiB7CiAgQG92ZXJyaWRlCiAgV2lkZ2V0IGJ1aWxkKEJ1aWxkQ29udGV4dCBjb250ZXh0KSB7CiAgICByZXR1cm4gU2NhZmZvbGQoKTsKICB9Cn0Ke3svc3RhdGV9fQoKe3tec3RhdGV9fQpjbGFzcyB7e25hbWUucGFzY2FsQ2FzZSgpfX1QYWdlIGV4dGVuZHMgU3RhdGVsZXNzV2lkZ2V0IHsKY29uc3Qge3tuYW1lLnBhc2NhbENhc2UoKX19UGFnZSh7c3VwZXIua2V5fSk7CgpAb3ZlcnJpZGUKV2lkZ2V0IGJ1aWxkKEJ1aWxkQ29udGV4dCBjb250ZXh0KSB7CnJldHVybiBTY2FmZm9sZCgpOwp9Cn0Ke3svc3RhdGV9fQ==",
+      "type": "text"
+    }
+  ],
+  "hooks": [
+    {
+      "path": "post_gen.dart",
+      "data":
+          "aW1wb3J0ICdkYXJ0OmlvJzsKCmltcG9ydCAncGFja2FnZTptYXNvbi9tYXNvbi5kYXJ0JzsKCnZvaWQgcnVuKEhvb2tDb250ZXh0IGNvbnRleHQpIHsKICBmaW5hbCBmaWxlID0gRmlsZSgne3tuYW1lLnNuYWtlQ2FzZSgpfX0uZGFydCcpOwogIGlmICghZmlsZS5leGlzdHNTeW5jKCkpIGZpbGUuY3JlYXRlU3luYygpOwogIGZpbGUud3JpdGVBc1N0cmluZ1N5bmMoCiAgICAgICdleHBvcnQgXCdwcmVzZW50YXRpb24vcGFnZXMve3tuYW1lLnNuYWtlQ2FzZSgpfX1fcGFnZS5kYXJ0XCc7XG4nLAogICAgICBtb2RlOiBGaWxlTW9kZS5hcHBlbmQpOwp9Cg==",
+      "type": "text"
+    },
+    {
+      "path": "pubspec.yaml",
+      "data":
+          "bmFtZTogY2xlYW5fZmVhdHVyZV9ob29rcwoKZW52aXJvbm1lbnQ6CiAgc2RrOiAiPj0yLjE3LjAgPDMuMC4wIgoKZGVwZW5kZW5jaWVzOgogIG1hc29uOiBhbnkK",
+      "type": "text"
+    }
+  ],
+  "name": "graphql_feature",
+  "description": "A new brick created with the Mason CLI.",
+  "version": "0.1.0+1",
+  "environment": {"mason": ">=0.1.0-dev <0.1.0"},
+  "readme": {
+    "path": "README.md",
+    "data":
+        "IyBjbGVhbl9mZWF0dXJlCgpBIG5ldyBicmljayBjcmVhdGVkIHdpdGggdGhlIE1hc29uIENMSS4KCl9HZW5lcmF0ZWQgYnkgW21hc29uXVsxXSDwn6exXwoKIyMgR2V0dGluZyBTdGFydGVkIPCfmoAKClRoaXMgaXMgYSBzdGFydGluZyBwb2ludCBmb3IgYSBuZXcgYnJpY2suCkEgZmV3IHJlc291cmNlcyB0byBnZXQgeW91IHN0YXJ0ZWQgaWYgdGhpcyBpcyB5b3VyIGZpcnN0IGJyaWNrIHRlbXBsYXRlOgoKLSBbT2ZmaWNpYWwgTWFzb24gRG9jdW1lbnRhdGlvbl1bMl0KLSBbQ29kZSBnZW5lcmF0aW9uIHdpdGggTWFzb24gQmxvZ11bM10KLSBbVmVyeSBHb29kIExpdmVzdHJlYW06IEZlbGl4IEFuZ2Vsb3YgRGVtb3MgTWFzb25dWzRdCgpbMV06IGh0dHBzOi8vZ2l0aHViLmNvbS9mZWxhbmdlbC9tYXNvbgpbMl06IGh0dHBzOi8vZ2l0aHViLmNvbS9mZWxhbmdlbC9tYXNvbi90cmVlL21hc3Rlci9wYWNrYWdlcy9tYXNvbl9jbGkjcmVhZG1lClszXTogaHR0cHM6Ly92ZXJ5Z29vZC52ZW50dXJlcy9ibG9nL2NvZGUtZ2VuZXJhdGlvbi13aXRoLW1hc29uCls0XTogaHR0cHM6Ly95b3V0dS5iZS9HNFBUakE2dHBUVQo=",
+    "type": "text"
+  },
+  "changelog": {
+    "path": "CHANGELOG.md",
+    "data": "IyAwLjEuMCsxCgotIFRPRE86IERlc2NyaWJlIGluaXRpYWwgcmVsZWFzZS4K",
+    "type": "text"
+  },
+  "license": {
+    "path": "LICENSE",
+    "data": "VE9ETzogQWRkIHlvdXIgbGljZW5zZSBoZXJlLgo=",
+    "type": "text"
+  },
+  "vars": {
+    "name": {
+      "type": "string",
+      "description": "Your feature name",
+      "prompt": "What is your feature name?"
+    },
+    "useDio": {
+      "type": "boolean",
+      "description": "client type",
+      "default": false,
+      "prompt": "use dio?"
+    },
+    "package_name": {
+      "type": "string",
+      "description": "Your package name",
+      "prompt": "What is your package name?"
+    },
+    "state": {
+      "type": "boolean",
+      "description": "Page type",
+      "default": false,
+      "prompt": "What is your page type <stateless,stateful>?"
+    }
+  }
+});